Tidying and slicing up to make more reusable
The use is currently very specific, with user-defined keys hardcoded in and so on. I'd like to make it more generic so it can be reused, so I'm slicing things into logical boxes and trying to replace copy-paste code with keyed loops. The keys can then be abstracted out into something like a command-line argument set or a config file.