Added JSON pukeproofing
I downloaded some JSON from BigQuery and it was of the form:

{"key":"value", "key2":"value2"}
{"key":"value", "key2":"value2"}

rather than

[
{"key":"value", "key2":"value2"},
{"key":"value", "key2":"value2"}
]

as any sane person would produce. What gives, Google? Anyway, this is a very basic fix to deal with JSON puking. It will simply refuse to handle invalid JSON but rather than crashing, it will continue and parse any valid JSON, leaving an error message on the console.

Longer term I will add an arrayenize function that makes JSON sensible again (if JSON is not sensible).